removed executor/executor.go
